requirements.txt
# Core ML Libraries
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0

# Additional utilities
Pillow>=10.0.0
